fixed realtime and moved socket to consts.js
